refactor(test): extract parser helpers to reduce command sprawl
